Feat: add bootstrap, prune, clean scripts
* Bootstrap.sh is added to handle installing theme dependencies as well
as setting the theme name in 2 different files as required by devloco's
wp theme script

* docker_prune.sh was added to simplify getting rid of the junk that
docker leaves behind when quitting. Currently this script does not do
anything to save the db data. In a future commit it can be ensured
that a db backup is made and save to ./backups

* theme_clean.sh clears all the junk files, node_modules and build
files.