Feat: implement custom file restore to db_restore
* It's possible that the user could want to restore a sql file that
isn't the latest. This comit allows restoring any sql through the use
of -f or --filename flags. If no flag is provided, the script restores
the latest sql like it did before

* Errors have been refactored as separate functions and moved to the
top of the file.

* Mysql output has been sent to /dev/null inside its container. this
may be a bad idea and could be replaced with a more selective silencing
algorithm later on