Feat: add replace url script
* Wordpress is aware of the url from which the site is being served.
Because of this, it's required to alter the "home" url while moving
db image between production and development.

This commit adds a script that is accesible from root/db_replace_url.sh
to alter the db home url from the cli.

This script can both be used to apply some prexisting db backup as well
as preparing the dev backup to be used in production