Chore: refactor db_backup.sh
* The control flow for checking whether the container is running inside
the container has been moved above the business logic.

This is done because there is no reason for the script to start any
business tasks before it's assured that the environment is correct.
Also, now the error message and the code that checks whether there
is an error is adjacent to each other.