Update: move root scripts to scripts/host
* root/wrt.sh file is now the entrypoint for all the scripts. This
is done to simplify the api for the repo

* All root scripts that were previously called directly from the
project root are now inside scripts/host folder. They can all be
called through  wrt.sh

* db_backup.sh had the db container name set without the __dev prefix.
This was fixed.

* theme_dist was added to create zipped distribution files to use
with wordpress.org or some direct upload to a cloud/ftp site.

* HOST_SCRIPTS variable was added to vars.sh. This value is used
extensively among shell scripts which need to know where they are
relative to wrt.

* db, theme and theme .sh files were added to simplify the api of wrt.sh