Update: refactor, fix multiple scripts
* db/replace_url messages have been moved to db/messsages.sh

* check_env was moved to folder scripts/shared

* logic for checking whether the code is running in a devcontainer was
refactored as a function and moved to shared. This script is now
referenced in other scripts that need to know which environment they
are running in,

* All the literary sections in scripts, such as errors and
notifications have been moved to shared/messages.sh. Scripts that run in
devcontainers are excluded from this change.

* argument parsing in scripts have been separated into multiple parts.
Essential arguments such as -h, --help and illegal command and option
checking has been abstracted from the rest of the argument loop and
saved as parse_args_essential. All arg loops now reference this
function.

Another function that only checks for these basic help and error args
have been created and referenced in all scripts that do not process any
arguments other than the basics. This new function is called
parse_args_basic.

* shared/vars.sh THEME_DIR has been renamed to THEME_SRC_DIR. This name
better explains the value that the variable is holding.

* Almost all script files received refactoring in which the script logic
was converted to a function.

* All script files have had their title and commands_and_options values
rewritten both for consistency and for readability of the files.

* Instructions of production placeholder scripts for pulling, pushing,
initialization have been rewritten to be more understandable