Update: add .env values check to check_env.sh
* Previously only the existence of the .env file was being checked. This
meant that there could have been missing properties in env and these
wouldn't have been noticed until something went wrong.

This update mitigates this issue for any script file that runs
check_env.

One notable exception here is devcontainer start script.
This script also needs to be updated to be aware that the values it
needs may not be set.