Update: add check_env in postAttachCommand
* devcontainer was starting react without checking whether the
environment variables were set. This could be a possible flow if the
user doesn't read the instructions on how to bootstrap the application.

With this commit, postAttachCommand produces some error, telling the
user that something is missing in the env file. But this error could
easily be overseen among a ton of other lines that vscode terminal
prints. There could be a better solution in the future but this
one is a start for telling the user that something is wrong with
their setup.