Feat: add check_ownership script
* This script will enable checking for whether the operation that is
about to begin can be handled without sudo.

The aim is to inform the user about ownership inconsistencies that
is a common issue while using docker.

It could have been possible to set a user in devcontainer settings
but letting docker do its thing forces the user to use
devcontainers rather than the host for developing